Attempt to develop an autoreformat hook
I would appreciate for people to try this locally before merging as this
is the first time I develop an recommit hook, so I'm not sure I did
everything right.

Note that precommit-hooks only run vélin on changed files so this is
complementary to the other PR that update some docstrings and may
conflicts as I made some heuristics updates (to ignore parameters named
'cls', and 'kwargs') so it found more typoes.